moved jetty rest implementations to rest package
