+++ commons-jetty/src/main/scala/com/avsystem/commons/jetty/rest/RestClient.scala
@@ -1,5 +1,5 @@
 package com.avsystem.commons
-package jetty.rpc
+package jetty.rest
 
 import java.net.URLEncoder
 import java.nio.charset.StandardCharsets
