+++ commons-jetty/src/test/scala/com/avsystem/commons/jetty/rpc/RestServletTest.scala
@@ -3,6 +3,7 @@ package jetty.rpc
 
 import java.nio.charset.StandardCharsets
 
+import com.avsystem.commons.jetty.rest.RestServlet
 import org.eclipse.jetty.client.util.StringContentProvider
 import org.eclipse.jetty.http.{HttpMethod, HttpStatus}
 import org.eclipse.jetty.server.Server
